Refactored WS Events to segregate them into projects. Events without a designated project will now route to the `events` channel, while project-specific events will be directed to `events.project.<key>` channels.
Implemented automatic subscription of users to relevant channels upon WS connection. Users will now be subscribed to both the general `events` channel and all channels pertaining to projects.

+++ app/modules/Events/Interfaces/Commands/DeleteEventHandler.php
@@ -14,15 +14,25 @@ final readonly class DeleteEventHandler
 {
     public function __construct(
         private EventRepositoryInterface $events,
-        private EventDispatcherInterface $dispatcher
+        private EventDispatcherInterface $dispatcher,
     ) {
     }
 
     #[CommandHandler]
     public function __invoke(DeleteEvent $command): void
     {
+        $event = $this->events->findByPK((string)$command->uuid);
+        if (!$event) {
+            return;
+        }
+
         if ($this->events->deleteByPK((string)$command->uuid)) {
-            $this->dispatcher->dispatch(new EventWasDeleted($command->uuid));
+            $this->dispatcher->dispatch(
+                new EventWasDeleted(
+                    uuid: $command->uuid,
+                    project: $event->getProject(),
+                ),
+            );
         }
     }
 }
